drg: refactor buffs to tsx, combine outputs
