+++ locale/de/messages.json
@@ -388,33 +388,33 @@
   "drg.blood.trigger.no-lotd-nastrond": "<0/> wurde benutzt während Drachengeist als inaktiv angesehen wurde.",
   "drg.blood.trigger.no-lotd-stardiver": "<0/> wurde benutzt während Drachengeist als inaktiv angesehen wurde.",
   "drg.blood.windows.hits": "",
-  "drg.buffs.accordion.ds-header": "Drachenauge",
-  "drg.buffs.accordion.lc-header": "",
   "drg.buffs.bad-start": "",
+  "drg.buffs.battle-lit.header": "",
   "drg.buffs.checklist.description": "",
   "drg.buffs.checklist.name": "Halte {0} aufrecht",
   "drg.buffs.checklist.requirement.name": "<0/> Gesamtanwendungszeit",
+  "drg.buffs.dragon-sight.header": "",
   "drg.buffs.gcd-count": "GCDs",
-  "drg.buffs.note.partial": "",
-  "drg.buffs.notes": "",
-  "drg.buffs.suggestions.bad-dss.content": "",
-  "drg.buffs.suggestions.bad-dss.why": "",
-  "drg.buffs.suggestions.bad-lcs.content": "",
-  "drg.buffs.suggestions.bad-lcs.why": "",
+  "drg.buffs.lance-charge.header": "",
   "drg.buffs.suggestions.life-surge.content": "",
   "drg.buffs.suggestions.life-surge.why": "",
   "drg.buffs.suggestions.solo-ds.content": "Obwohl es deinen persönlichen DPS nicht beeinflusst, solltest du Drachenblick immer auf einen Partner in deiner Gruppe benutzen, damit auch andere von dem Schadensbonus profitieren können.",
   "drg.buffs.suggestions.solo-ds.why": "",
   "drg.buffs.title": "",
+  "drg.buffwindow.suggestions.bad-gcd": "",
+  "drg.buffwindow.suggestions.bad-gcd.why": "",
+  "drg.buffwindow.title": "",
   "drg.debuffs.checklist.description": "",
   "drg.debuffs.checklist.name": "Halte deine DoTs aufrecht",
   "drg.debuffs.checklist.requirement.chaos-thrust.name": "Gesamtanwendungszeit von <0/>",
   "drg.debuffs.suggestions.clipping.content": "",
   "drg.debuffs.suggestions.clipping.why": "",
+  "drg.dragonsight.title": "",
   "drg.drift.table.heading-1": "",
   "drg.drift.table.heading-2": "",
   "drg.drift.table.heading-3": "Gesamter Drift",
   "drg.drift.title": "",
+  "drg.lancecharge.title": "",
   "drg.positionals.analysis.missed": "",
   "drg.positionals.analysis.truenorth": "",
   "drg.positionals.suggestions.content": "",
