+++ locale/en/messages.json
@@ -388,33 +388,33 @@
   "drg.blood.trigger.no-lotd-nastrond": "<0/> was cast while Life of the Dragon was deemed inactive.",
   "drg.blood.trigger.no-lotd-stardiver": "<0/> was cast while Life of the Dragon was deemed inactive.",
   "drg.blood.windows.hits": "{0, plural, one {# Nastrond} other {# Nastronds}}, {1, plural, one {# Stardiver} other {# Stardivers}}",
-  "drg.buffs.accordion.ds-header": "Dragon Sight",
-  "drg.buffs.accordion.lc-header": "Lance Charge",
   "drg.buffs.bad-start": "Optimal Start",
+  "drg.buffs.battle-lit.header": "Battle Litany",
   "drg.buffs.checklist.description": "<0/> provides a 10% boost to your personal damage and should always be kept up.",
   "drg.buffs.checklist.name": "Keep {0} up",
   "drg.buffs.checklist.requirement.name": "<0/> uptime",
+  "drg.buffs.dragon-sight.header": "Dragon Sight",
   "drg.buffs.gcd-count": "GCDs",
-  "drg.buffs.note.partial": "Partial Window",
-  "drg.buffs.notes": "Notes",
-  "drg.buffs.suggestions.bad-dss.content": "Avoid using <0/> immediately after <1/> or <2/> in order to get the most possible damage out of each window.",
-  "drg.buffs.suggestions.bad-dss.why": "{badDragonSights} of your Dragon Sight windows started right after a standard combo finisher.",
-  "drg.buffs.suggestions.bad-lcs.content": "Avoid using <0/> immediately after <1/> or <2/> in order to get the most possible damage out of each window.",
-  "drg.buffs.suggestions.bad-lcs.why": "{badLanceCharges} of your Lance Charge windows started right after a standard combo finisher.",
+  "drg.buffs.lance-charge.header": "Lance Charge",
   "drg.buffs.suggestions.life-surge.content": "Avoid using <0/> on any GCD that isn't <1/> or a 5th combo hit. Any other combo action will have significantly less potency, losing a lot of the benefit of the guaranteed crit.",
   "drg.buffs.suggestions.life-surge.why": "You used {0} on a non-optimal GCD {1, plural, one {# time} other {# times}}.",
   "drg.buffs.suggestions.solo-ds.content": "Although it doesn't impact your personal DPS, try to always use Dragon Sight on a partner in group content so that someone else can benefit from the damage bonus too.",
   "drg.buffs.suggestions.solo-ds.why": "At least 1 of your Dragon Sight casts didn't have a tether partner.",
-  "drg.buffs.title": "Lance Charge & Dragon Sight",
+  "drg.buffs.title": "Buffs",
+  "drg.buffwindow.suggestions.bad-gcd": "Avoid using <0/> immediately after <1/> or <2/> in order to get the most possible damage out of each window.",
+  "drg.buffwindow.suggestions.bad-gcd.why": "{badBuffs} of your <0/> windows started right after a standard combo finisher.",
+  "drg.buffwindow.title": "Dragoon Buff Window",
   "drg.debuffs.checklist.description": "<0/> provides a potent DoT which should be maintained at all times.",
   "drg.debuffs.checklist.name": "Keep your debuffs up",
   "drg.debuffs.checklist.requirement.chaos-thrust.name": "<0/> uptime",
   "drg.debuffs.suggestions.clipping.content": "Avoid refreshing <0/> significantly earlier or later than its expiration, as it usually indicates rotational errors. DRG's strict 10-GCD rotation should have you refreshing Chaos Thrust within 2 seconds before or after expiry, depending on your skill speed.",
   "drg.debuffs.suggestions.clipping.why": "You lost {0} of Chaos Thrust to early refreshes.",
+  "drg.dragonsight.title": "Dragon Sight",
   "drg.drift.table.heading-1": "Casts",
   "drg.drift.table.heading-2": "Drift",
   "drg.drift.table.heading-3": "Total Drift",
   "drg.drift.title": "Ability Drift",
+  "drg.lancecharge.title": "Lance Charge",
   "drg.positionals.analysis.missed": "<0/> You missed <1>{missed}</1> of <2>{0}</2> possible <3/> procs.",
   "drg.positionals.analysis.truenorth": "Of these missed procs, <0>{withTn}</0> could be handled with <1/>.",
   "drg.positionals.suggestions.content": "Performing the proper positionals for <0/> and <1/> will do more damage and let you use <2/>. Try to land these positionals, and use <3/> if it's available.",
