+++ locale/fr/messages.json
@@ -388,33 +388,33 @@
   "drg.blood.trigger.no-lotd-nastrond": "<0/> a été utiliser alors que vie du dragon était inactif.",
   "drg.blood.trigger.no-lotd-stardiver": "<0/> a été utiliser alors que vie du dragon était inactif.",
   "drg.blood.windows.hits": "{0, plural,one {Naströnd}other {#Naströnds}},{1, plural,one {#Plongeon céleste}other {#Plongeons célestes}}",
-  "drg.buffs.accordion.ds-header": "Vue de dragon",
-  "drg.buffs.accordion.lc-header": "Lance acérée",
   "drg.buffs.bad-start": "Départ ok?",
+  "drg.buffs.battle-lit.header": "",
   "drg.buffs.checklist.description": "<0/> vous octroie 10% de dégâts en plus et devrait être actif à tout moment.",
   "drg.buffs.checklist.name": "Gardez {0} actif",
   "drg.buffs.checklist.requirement.name": "Temps d'activité total de <0/>",
+  "drg.buffs.dragon-sight.header": "",
   "drg.buffs.gcd-count": "GCDs",
-  "drg.buffs.note.partial": "Fenêtre partielle",
-  "drg.buffs.notes": "Notes",
-  "drg.buffs.suggestions.bad-dss.content": "Evitez d'utiliser <0/> immédiatement après <1/> ou <2/> afin d'optimiser vos dégâts durant chaque fenêtre de buff.",
-  "drg.buffs.suggestions.bad-dss.why": "{badDragonSights} de votre fenêtre de Vue de dragon démarre directement après un combos final standard.",
-  "drg.buffs.suggestions.bad-lcs.content": "Evitez d'utiliser <0/> directement apres <1/> ou <2/> dans le but d'avoir le plus grand potentiel de dégat par phase de dps.",
-  "drg.buffs.suggestions.bad-lcs.why": "{badLanceCharges} de votre fenêtre de Lance acérée démarre directement après un combos final standard.",
+  "drg.buffs.lance-charge.header": "",
   "drg.buffs.suggestions.life-surge.content": "Évitez d'utiliser <0/> sur un GCD qui n'est ni <1/> ni la 5ème action de votre combo. Toute autre action de combo aura une puissance significativement inférieure, vous faisant perdre une grande partie du bénéfice d'un coup critique garanti.",
   "drg.buffs.suggestions.life-surge.why": "Vous avez utilisé {0} sur un GCD non optimal {1, plural, one {# fois} other {# fois}}.",
   "drg.buffs.suggestions.solo-ds.content": "Bien que cela n'ait pas d'impact sur votre DPS personnel, essayez toujours d'utiliser vue du dragon sur un partenaire dans le groupe afin que quelqu'un d'autre puisse également bénéficier du bonus de dommages.",
   "drg.buffs.suggestions.solo-ds.why": "Au moins un de vos vue du dragon n'avait pas de partenaire.",
   "drg.buffs.title": "Lance acérée et Vue de dragon",
+  "drg.buffwindow.suggestions.bad-gcd": "",
+  "drg.buffwindow.suggestions.bad-gcd.why": "",
+  "drg.buffwindow.title": "",
   "drg.debuffs.checklist.description": "<0/> fournit un DoT puissant qui doit être maintenu en tout temps.",
   "drg.debuffs.checklist.name": "Maintenez vos débuffs actifs",
   "drg.debuffs.checklist.requirement.chaos-thrust.name": "Temps d'activité total de <0/>",
   "drg.debuffs.suggestions.clipping.content": "Évitez de rafraîchir <0/> trop tôt avant son expiration. En règle générale, cela indique une erreur dans votre rotation. Sachant que la rotation du DRG comporte 10 GCDs, vous devriez ré-appliquer l'effet de Percée Chaotique que lorsqu'il ne reste 1 ou 2 secondes avant son expiration (selon votre vivacité).",
   "drg.debuffs.suggestions.clipping.why": "Vous avez perdu {0} utilisations de Percée Chaotique à cause de ré-applications précipitées.",
+  "drg.dragonsight.title": "",
   "drg.drift.table.heading-1": "Sorts",
   "drg.drift.table.heading-2": "Delta",
   "drg.drift.table.heading-3": "Delta Total",
   "drg.drift.title": "",
+  "drg.lancecharge.title": "",
   "drg.positionals.analysis.missed": "<0/> vous avez manqué<1>{missed}</1> sur <2>{0}</2> possible <3/> procs.",
   "drg.positionals.analysis.truenorth": "De ces procs manqués, <0>{withTn}</0> pouvait être corrigé avec <1/>.",
   "drg.positionals.suggestions.content": "Effectuer les positions appropriées pour <0/> et <1/> fera plus de dégâts et vous permettra d'utiliser <2/>. Essayez d'utiliser ces positionnelles, et utilisez <3/> s'il est disponible.",
