+++ locale/ja/messages.json
@@ -388,33 +388,33 @@
   "drg.blood.trigger.no-lotd-nastrond": "紅の竜血が無いと思われるときに<0/>が使用されました。",
   "drg.blood.trigger.no-lotd-stardiver": "紅の竜血が無いと思われるときに<0/>が使用されました。",
   "drg.blood.windows.hits": "{0, plural, other {# 回のナーストレンド}}, {1, plural, other {# 回のスターダイバー}}",
-  "drg.buffs.accordion.ds-header": "ドラゴンサイト",
-  "drg.buffs.accordion.lc-header": "ランスチャージ",
   "drg.buffs.bad-start": "使用タイミング",
+  "drg.buffs.battle-lit.header": "",
   "drg.buffs.checklist.description": "<0/> は与ダメージ10％アップで常に維持すべきです。",
   "drg.buffs.checklist.name": "{0} の維持率",
   "drg.buffs.checklist.requirement.name": "アクティブ率: <0/>",
+  "drg.buffs.dragon-sight.header": "",
   "drg.buffs.gcd-count": "GCDs",
-  "drg.buffs.note.partial": "効果時間の一部",
-  "drg.buffs.notes": "ノート",
-  "drg.buffs.suggestions.bad-dss.content": "効果時間中に最大限ダメージを与えるために、<0/>は<1/>や<2/>の直後では使用しないでください。",
-  "drg.buffs.suggestions.bad-dss.why": "{badDragonSights} 回ドラゴンサイトをコンボの直後に使用した",
-  "drg.buffs.suggestions.bad-lcs.content": "効果時間中に最大限ダメージを与えるために、<0/>は<1/>や<2/>の直後では使用しないでください。",
-  "drg.buffs.suggestions.bad-lcs.why": "{badLanceCharges} 回ランスチャージをコンボの直後に使用した",
+  "drg.buffs.lance-charge.header": "",
   "drg.buffs.suggestions.life-surge.content": "<0/> は<1/>かコンボの5段目に使用するようにしましょう。他の箇所では確定クリティカルの恩恵が小さくなります。",
   "drg.buffs.suggestions.life-surge.why": "{1, plural, other {# 回}} の{0} が適さないウェポンスキルに対して使用されました。",
   "drg.buffs.suggestions.solo-ds.content": "あなた個人のDPSには影響しませんが、パーティーメンバーのだれかがダメージボーナスを得られるように、常にドラゴンサイトを使うようにしてみてください。",
   "drg.buffs.suggestions.solo-ds.why": "最低でも1回のドラゴンサイトに対象のメンバーが指定されませんでした。",
   "drg.buffs.title": "ランスチャージ＆ドラゴンサイト",
+  "drg.buffwindow.suggestions.bad-gcd": "",
+  "drg.buffwindow.suggestions.bad-gcd.why": "",
+  "drg.buffwindow.title": "",
   "drg.debuffs.checklist.description": "<0/> は強力なDoTで、常に維持されるべきです。",
   "drg.debuffs.checklist.name": "デバフを常に切らさないようにしてください。",
   "drg.debuffs.checklist.requirement.chaos-thrust.name": "アクティブ率: <0/>",
   "drg.debuffs.suggestions.clipping.content": "<0/>の効果が満了するまでの早すぎる更新は避けましょう。普通はスキル回しの間違いです。竜騎士は10GCDのスキル回しが決まっており、あなたのスキルスピードに応じて1～2秒残しで桜花狂咲を更新しましょう。",
   "drg.debuffs.suggestions.clipping.why": "早すぎる更新により桜華狂咲を失った回数: {0} 回",
+  "drg.dragonsight.title": "",
   "drg.drift.table.heading-1": "",
   "drg.drift.table.heading-2": "遅延",
   "drg.drift.table.heading-3": "更新遅延総計",
   "drg.drift.title": "",
+  "drg.lancecharge.title": "",
   "drg.positionals.analysis.missed": "<0/> あなたは<2>{0}</2>回中<1>{missed}</1>回の<3/>プロックをミスしました。",
   "drg.positionals.analysis.truenorth": "ミスしたプロックのうち、<0>{withTn}</0>回は<1/>が使用可能でした。",
   "drg.positionals.suggestions.content": "<0/>と<1/>の方向指定を正しくとれば、与えるダメージが増え<2/>が使えるようになります。使用可能なら<3/>も使い、これらの方向指定を取れるようにしましょう。",
