+++ locale/ko/messages.json
@@ -388,33 +388,33 @@
   "drg.blood.trigger.no-lotd-nastrond": "붉은 용혈이 없는 동안 <0/>를 사용했습니다.",
   "drg.blood.trigger.no-lotd-stardiver": "붉은 용혈이 없는 동안 <0/>를 사용했습니다.",
   "drg.blood.windows.hits": "나스트론드 {0, plural, other {#회}}, 스타 다이버 {1, plural, other {#회}} 사용",
-  "drg.buffs.accordion.ds-header": "용의 시선",
-  "drg.buffs.accordion.lc-header": "돌격하는 창",
   "drg.buffs.bad-start": "올바르게 시작하셧습니까?",
+  "drg.buffs.battle-lit.header": "",
   "drg.buffs.checklist.description": "<0/>는 피해량을 10% 증가시키며 항상 켜져 있어야 합니다.",
   "drg.buffs.checklist.name": "항상 {0}가 유지되도록 하십시오.",
   "drg.buffs.checklist.requirement.name": "적용시간 <0/>",
+  "drg.buffs.dragon-sight.header": "",
   "drg.buffs.gcd-count": "글쿨기",
-  "drg.buffs.note.partial": "쪼개진 구간",
-  "drg.buffs.notes": "노트",
-  "drg.buffs.suggestions.bad-dss.content": "<0/>을 <1/> 혹은 <2/> 직후 사용하는것을 피하여 딜을 최대한 늘리십시오.",
-  "drg.buffs.suggestions.bad-dss.why": "용의 시선을 기본 콤보를 끝낸 이후 사용한 실수를 총 {badDragonSights} 회 하였습니다.",
-  "drg.buffs.suggestions.bad-lcs.content": "<0/>을 <1/> 혹은 <2/> 직후 사용하는것을 피하여 딜을 최대한 늘리십시오.",
-  "drg.buffs.suggestions.bad-lcs.why": "돌격하는 창을 기본 콤보를 끝낸 이후 사용한 실수를 총 {badLanceCharges} 회 하였습니다.",
+  "drg.buffs.lance-charge.header": "",
   "drg.buffs.suggestions.life-surge.content": "<0/>를 <1/>나 5번째 콤보가 아닌 글쿨에 사용하는 것을 피하세요. 다른 무기 기술은 현저하게 낮은 위력으로 인해 그 효과가 떨어집니다.",
   "drg.buffs.suggestions.life-surge.why": "{0} 를 적절하지 못한 무기 기술에 {0, plural, one {#회} other {#회}} 사용했습니다.",
   "drg.buffs.suggestions.solo-ds.content": "대상을 지정하지 않은 용의 시선은 파티원에게 용의 왼눈 버프를 주지 못합니다. 개인 딜에 영향이 안끼치더라도 최대한 대상을 지정하여 사용하십시오.",
   "drg.buffs.suggestions.solo-ds.why": "대상을 지정하지 않은 용의 시선이 최소 1 회 이상 사용되었습니다.",
   "drg.buffs.title": "돌격하는 창 & 용의 눈",
+  "drg.buffwindow.suggestions.bad-gcd": "",
+  "drg.buffwindow.suggestions.bad-gcd.why": "",
+  "drg.buffwindow.title": "",
   "drg.debuffs.checklist.description": "<0/>는 상당한 지속 데미지를 주며 상시 유지시켜야 합니다.",
   "drg.debuffs.checklist.name": "항상 디버프가 유지되도록 하십시오",
   "drg.debuffs.checklist.requirement.chaos-thrust.name": "<0/> 지속시간",
   "drg.debuffs.suggestions.clipping.content": "<0/>이 만료되기전에 갱신됬다는 것은 딜사이클이 꼬인 증거입니다. 용기사의 10글쿨 딜사이클은 매우 엄격하며 기술 시전 속도에 따라 꽃잎 폭풍이 1-2초를 남긴 상태에서 갱신됩니다.",
   "drg.debuffs.suggestions.clipping.why": "너무 이르게 갱신하여 {0} 만큼의 꽃잎 폭풍 지속시간을 잃었습니다.",
+  "drg.dragonsight.title": "",
   "drg.drift.table.heading-1": "시전 기술",
   "drg.drift.table.heading-2": "지연된 시간",
   "drg.drift.table.heading-3": "총 갱신지연된 시간",
   "drg.drift.title": "스킬 지연",
+  "drg.lancecharge.title": "",
   "drg.positionals.analysis.missed": "<0/> <3/>프록의 총 <2>{0}</2>회중 <1>{missed}</1> 회가 빗나갔습니다",
   "drg.positionals.analysis.truenorth": "<1/>으로 프록을 받을 수 있는 기회를 <0>{withTn}</0> 회 놓쳤습니다.",
   "drg.positionals.suggestions.content": "올바른 측후 보너스를 받은 <0/>과 <1/>는 더 높은 데미지와 함께 <2/> 사용을 가능케 합니다. 측후 보너스를 최대한 노리거나 <3/>을 사용하십시오.",
