+++ locale/zh/messages.json
@@ -388,33 +388,33 @@
   "drg.blood.trigger.no-lotd-nastrond": "红莲龙血尚未激活的时候使用了<0/>。",
   "drg.blood.trigger.no-lotd-stardiver": "红莲龙血尚未激活的时候使用了<0/>。",
   "drg.blood.windows.hits": "{0, plural, one {# 死者之岸} other {# 死者之岸}}，{1, plural, one {# 坠星冲} other {# 坠星冲}}",
-  "drg.buffs.accordion.ds-header": "巨龙视线",
-  "drg.buffs.accordion.lc-header": "猛枪",
   "drg.buffs.bad-start": "使用时间",
+  "drg.buffs.battle-lit.header": "",
   "drg.buffs.checklist.description": "<0/>可以提高自身伤害10%，应该始终保持。",
   "drg.buffs.checklist.name": "保持 {0} 不断",
   "drg.buffs.checklist.requirement.name": "<0/> 覆盖率",
+  "drg.buffs.dragon-sight.header": "",
   "drg.buffs.gcd-count": "GCD技能",
-  "drg.buffs.note.partial": "一部分效果持续时间",
-  "drg.buffs.notes": "备注",
-  "drg.buffs.suggestions.bad-dss.content": "不要在<1/>或<2/>后立刻就用<0/>以使威力最大化。",
-  "drg.buffs.suggestions.bad-dss.why": "你在标准连击的末尾（直刺/樱花怒放）使用了{badDragonSights} 个巨龙视线",
-  "drg.buffs.suggestions.bad-lcs.content": "不要在<1/>或<2/>后立刻就用<0/>以使威力最大化。",
-  "drg.buffs.suggestions.bad-lcs.why": "你在标准连击的末尾（直刺/樱花怒放）使用了{badLanceCharges} 个猛枪",
+  "drg.buffs.lance-charge.header": "",
   "drg.buffs.suggestions.life-surge.content": "不要在<1/>与龙5以外的GCD使用<0/>。用在其他任何连击技能上，都会显著降低威力，损失大量必定暴击带来的收益。",
   "drg.buffs.suggestions.life-surge.why": "你在非最优GCD使用了 {1, plural, other {# 次}}{0}",
   "drg.buffs.suggestions.solo-ds.content": "尽管不连队友不会影响你的个人dps，但它是你的重要的团辅，不连就白白浪费掉了，随便找个可以连上的人都好。",
   "drg.buffs.suggestions.solo-ds.why": "至少有一次巨龙视线没有绑定队友。",
   "drg.buffs.title": "猛枪&巨龙视线",
+  "drg.buffwindow.suggestions.bad-gcd": "",
+  "drg.buffwindow.suggestions.bad-gcd.why": "",
+  "drg.buffwindow.title": "",
   "drg.debuffs.checklist.description": "<0/>提供强力的DoT伤害所以请保持常驻。",
   "drg.debuffs.checklist.name": "维持你的Debuff覆盖",
   "drg.debuffs.checklist.requirement.chaos-thrust.name": "<0/> 覆盖率",
   "drg.debuffs.suggestions.clipping.content": "不要过早地刷新<0/>，这样通常意味着循环错误。龙骑士严格的10-GCD循环中樱花怒放应该会在还剩1~2秒的时候刷新，具体时间取决于你的技能速度。",
   "drg.debuffs.suggestions.clipping.why": "由于过早的刷新，你损失了 {0} 的樱花怒放 。",
+  "drg.dragonsight.title": "",
   "drg.drift.table.heading-1": "咏唱",
   "drg.drift.table.heading-2": "推移",
   "drg.drift.table.heading-3": "总推移时间",
   "drg.drift.title": "能力技偏移",
+  "drg.lancecharge.title": "",
   "drg.positionals.analysis.missed": "<0/> 在可能的<2>{0}</2>次<3/>触发中，你丢失了<1>{missed}</1>次。",
   "drg.positionals.analysis.truenorth": "在你失误的触发类技能中，你可以多打<0>{withTn}</0>次<1/>。",
   "drg.positionals.suggestions.content": "以正确的身位打出<0/>和<1/>，会让你在打出更高伤害的同时，可在下一次循环中使用<2/>。尝试在这两个连击中打出正确的身位，或在可用的时候使用<3/>。",
